+++ src/main/java/org/example/learningsystem/lesson/dto/LessonResponseDto.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 import java.time.Instant;
 import java.util.UUID;
 
-import static org.example.learningsystem.core.util.format.DataFormatUtils.DATE_TIME_FORMAT;
+import static org.example.learningsystem.core.util.format.DateFormatUtils.DATE_TIME_FORMAT;
 
 @Schema
 @Getter
