+++ src/main/java/org/example/learningsystem/student/dto/StudentResponseDto.java
@@ -9,8 +9,8 @@ import java.time.LocalDate;
 import java.util.Locale;
 import java.util.UUID;
 
-import static org.example.learningsystem.core.util.format.DataFormatUtils.DATE_FORMAT;
-import static org.example.learningsystem.core.util.format.DataFormatUtils.DATE_TIME_FORMAT;
+import static org.example.learningsystem.core.util.format.DateFormatUtils.DATE_FORMAT;
+import static org.example.learningsystem.core.util.format.DateFormatUtils.DATE_TIME_FORMAT;
 
 @Schema
 public record StudentResponseDto(
