+++ src/test/integration/org/example/learningsystem/course/builder/CourseRequestBuilder.java
@@ -1,4 +1,4 @@
-package org.example.learningsystem.common.builder;
+package org.example.learningsystem.course.builder;
 
 import org.springframework.data.domain.PageRequest;
 import org.springframework.test.web.servlet.RequestBuilder;
@@ -12,32 +12,44 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 
-public class StudentRequestBuilder {
+public class CourseRequestBuilder {
 
-    private static final String STUDENTS_URL = "/students";
-    private static final String STUDENT_URL = "/students/{id}";
+    private static final String COURSES_URL = "/courses";
+    private static final String COURSE_URL = "/courses/{courseId}";
+    private static final String COURSE_LESSONS_URL = "/courses/{courseId}/lessons";
 
     public static RequestBuilder buildCreateRequest(String content) {
-        return post(STUDENTS_URL)
+        return post(COURSES_URL)
+                .contentType(APPLICATION_JSON)
+                .content(content);
+    }
+
+    public static RequestBuilder buildCreateLessonRequest(UUID courseId, String content) {
+        return post(COURSE_LESSONS_URL, courseId)
                 .contentType(APPLICATION_JSON)
                 .content(content);
     }
 
     public static RequestBuilder buildDeleteByIdRequest(UUID id) {
-        return delete(STUDENT_URL, id);
+        return delete(COURSE_URL, id);
     }
 
     public static RequestBuilder buildGetAllRequest(PageRequest pageRequest) {
-        var builder = get(STUDENTS_URL);
+        var builder = get(COURSES_URL);
         return build(builder, pageRequest);
     }
 
     public static RequestBuilder buildGetByIdRequest(UUID id) {
-        return get(STUDENT_URL, id);
+        return get(COURSE_URL, id);
+    }
+
+    public static RequestBuilder buildGetLessonsByIdRequest(UUID courseId, PageRequest pageRequest) {
+        var builder = get(COURSE_LESSONS_URL, courseId);
+        return build(builder, pageRequest);
     }
 
     public static RequestBuilder buildUpdateByIdRequest(UUID id, String content) {
-        return put(STUDENT_URL, id)
+        return put(COURSE_URL, id)
                 .contentType(APPLICATION_JSON)
                 .content(content);
     }
